Move photo greeting to photo activity
