+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".GreetingFormatActivity"></activity>
+        <activity android:name=".FormatViews.PhotoActivity"></activity>
+        <activity android:name=".GreetingFormatActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
