+++ app/src/main/res/layout/activity_greeting_format.xml
@@ -28,5 +28,6 @@
         android:layout_marginEnd="212dp"
         android:layout_marginRight="212dp"
         android:layout_marginBottom="164dp"
+        android:onClick="clickPhoto"
         app:srcCompat="@drawable/camera" />
 </RelativeLayout>
